[chore] #8 중복된 dependencies 삭제 및 패키지 구조 수정
