+++ app/src/main/res/layout/fragment_product_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.productDetail.ProductDetailFragment">
+    tools:context=".presentation.ui.productdetail.ProductDetailFragment">
 
     <TextView
         android:layout_width="wrap_content"
