+++ app/src/main/res/layout/fragment_recommend.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.main.recommend.RecommendFragment">
+    tools:context=".presentation.ui.main.home.recommend.RecommendFragment">
 
     <TextView
         android:layout_width="wrap_content"
