+++ app/src/main/res/layout/fragment_release.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.main.release.ReleaseFragment">
+    tools:context=".presentation.ui.main.home.release.ReleaseFragment">
 
     <androidx.compose.ui.platform.ComposeView
         android:id="@+id/cv_release"
